refactor paths constants
