+++ api/routes/jobs.py
@@ -8,14 +8,8 @@ import uuid
 from api.errors import ErrorCode, http_error
 from api.models import Job, JobStatusEnum, TranscriptMetadata
 from api.orm_bootstrap import SessionLocal
-from api.app_state import LOCAL_TZ
-from api.app_state import (
-    UPLOAD_DIR,
-    TRANSCRIPTS_DIR,
-    LOG_DIR,
-    db_lock,
-    handle_whisper,
-)
+from api.app_state import LOCAL_TZ, db_lock, handle_whisper
+from api.paths import UPLOAD_DIR, TRANSCRIPTS_DIR, LOG_DIR
 
 router = APIRouter()
