+++ api/routes/logs.py
@@ -2,7 +2,8 @@ from fastapi import APIRouter, Request
 from fastapi.responses import PlainTextResponse
 
 from api.errors import ErrorCode, http_error
-from api.app_state import LOG_DIR, ACCESS_LOG, backend_log
+from api.paths import LOG_DIR
+from api.app_state import ACCESS_LOG, backend_log
 
 router = APIRouter()
